# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NAEOS API
  description: REST API for NAEOS - Nusantara Engineering & Architecture Operating System
  version: 0.4.0
  contact:
    name: NAEOS Foundation
    url: https://github.com/NAEOS-foundation/naeos
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /specs:
    get:
      summary: List specifications
      tags: [Specifications]
      responses:
        '200':
          description: List of specifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecListResponse'
    post:
      summary: Upload specification
      tags: [Specifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specification'
      responses:
        '201':
          description: Specification received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecUploadResponse'
  /specs/validate:
    post:
      summary: Validate specification
      tags: [Specifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specification'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /specs/compile:
    post:
      summary: Compile specification to AI instruction sets
      tags: [Specifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specification'
      responses:
        '200':
          description: Compilation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileResponse'
...
components:
  schemas:
...
    Specification:
      type: object
      properties:
        project:
          type: string
        modules:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              path:
                type: string
        services:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              kind:
                type: string
              port:
                type: integer
    SpecListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            specs:
              type: array
              items:
                type: string
    SpecUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            message:
              type: string
            modules:
              type: integer
    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            valid:
              type: boolean
            errors:
              type: array
              items:
                type: string
            warnings:
              type: array
              items:
                type: string
    CompileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            compiled:
              type: boolean
            targets:
              type: array
              items:
                type: string
...